move media logic into controller
